removing idea files
