+++ .gitignore
@@ -166,3 +166,5 @@ gradle-app.setting
 # gradle/wrapper/gradle-wrapper.properties
 
 # End of https://www.gitignore.io/api/gradle,grails,node,intellij,visualstudiocode
+
+.idea
