--- .idea/artifacts/Gradle___com_openclicker___open_clicker_0_1_war.xml
@@ -1,11 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="war" name="Gradle : com.openclicker : open-clicker-0.1.war">
-    <output-path>$PROJECT_DIR$/build/libs</output-path>
-    <properties id="gradle-properties">
-      <options external-project-path="$PROJECT_DIR$" />
-    </properties>
-    <root id="archive" name="open-clicker-0.1.war">
-      <element id="artifact" artifact-name="Gradle : com.openclicker : open-clicker-0.1.war (exploded)" />
-    </root>
-  </artifact>
-</component>
