--- .idea/compiler.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <bytecodeTargetLevel>
-      <module name="open-clicker_integrationTest" target="1.8" />
-      <module name="open-clicker_main" target="1.8" />
-      <module name="open-clicker_test" target="1.8" />
-    </bytecodeTargetLevel>
-  </component>
-</project>
