+++ build.gradle
@@ -11,7 +11,7 @@ buildscript {
 }
 
 version "0.1"
-group "com.openclicker"
+group "com.lakerpolling"
 
 apply plugin:"eclipse"
 apply plugin:"idea"
@@ -46,17 +46,17 @@ dependencies {
     compile "org.hibernate:hibernate-core:5.1.2.Final"
     compile "org.hibernate:hibernate-ehcache:5.1.2.Final"
     compile 'com.google.api-client:google-api-client:1.22.0'
-    compile group: 'com.google.http-client', name: 'google-http-client-jackson', version: '1.22.0'
-    compile group: 'org.codehaus.groovy.modules.http-builder', name: 'http-builder', version: '0.7.1'
-    compile 'org.grails.plugins:jasypt-encryption:2.0.2'
+    compile  'com.google.http-client:google-http-client-jackson:1.22.0'
+    compile 'org.codehaus.groovy.modules.http-builder:http-builder:0.7.1'
+//    compile 'org.grails.plugins:jasypt-encryption:2.0.2'
 
     console "org.grails:grails-console"
     profile "org.grails.profiles:web"
 
-//    runtime 'mysql:mysql-connector-java:5.1.36'
+    runtime 'mysql:mysql-connector-java:5.1.36'
     runtime "com.bertramlabs.plugins:asset-pipeline-grails:2.11.6"
     runtime "com.h2database:h2"
-    runtime group: 'org.postgresql', name: 'postgresql', version: '9.4.1212.jre7'
+//    runtime group: 'org.postgresql', name: 'postgresql', version: '9.4.1212.jre7'
 
     testCompile "org.grails:grails-plugin-testing"
     testCompile "org.grails.plugins:geb"
