--- grails-app/controllers/com/openclicker/controllers/LogoutController.groovy
@@ -1,20 +1,0 @@
-package com.openclicker.controllers
-
-import grails.converters.JSON
-
-class LogoutController {
-
-    def authenticatorService
-
-    /**
-     * Endpoint to invalidate the server's tracking of this session.
-     */
-    def invalidateSignin() {
-        authenticatorService.signOutCurrentUser(session)
-        render ([
-                "success" : "true",
-                "message" : "ok"
-        ] as JSON)
-    }
-
-}
