+++ grails-app/domain/com/lakerpolling/models/authentication/OAuthId.groovy
@@ -1,0 +1,16 @@
+package com.lakerpolling.models.authentication
+
+class OAuthId  {
+
+    String subject
+
+    static belongsTo = [user : User]
+
+    static mapping = {
+        subject unique: true
+        autoTimestamp false
+        version false
+    }
+
+
+}
