--- grails-app/domain/com/openclicker/models/authentication/OAuthId.groovy
@@ -1,16 +1,0 @@
-package com.openclicker.models.authentication
-
-class OAuthId  {
-
-    String subject
-
-    static belongsTo = [user : User]
-
-    static mapping = {
-        subject unique: true
-        autoTimestamp false
-        version false
-    }
-
-
-}
