--- grails-app/domain/com/openclicker/models/authentication/User.groovy
@@ -1,21 +1,0 @@
-package com.openclicker.models.authentication
-
-class User {
-
-    String firstName
-    String lastName
-    String email
-    String imageUrl
-    String subject
-
-    static mapping = {
-        version false
-    }
-
-    static constraints = {
-        email email: true, unique: true
-        firstName blank: false
-        lastName blank:false
-    }
-
-}
