+++ grails-app/init/com/lakerpolling/Application.groovy
@@ -1,0 +1,10 @@
+package com.lakerpolling
+
+import grails.boot.GrailsApp
+import grails.boot.config.GrailsAutoConfiguration
+
+class Application extends GrailsAutoConfiguration {
+    static void main(String[] args) {
+        GrailsApp.run(Application, args)
+    }
+}
