+++ grails-app/init/com/lakerpolling/BootStrap.groovy
@@ -1,0 +1,16 @@
+package com.lakerpolling
+
+import com.lakerpolling.models.authentication.User
+
+class BootStrap {
+
+    def init = { servletContext ->
+
+    }
+
+
+    def destroy = {
+
+    }
+
+}
