--- src/main/groovy/com/clickerapi/util/Pair.groovy
@@ -1,14 +1,0 @@
-package com.clickerapi.util
-
-
-class Pair<K, V> {
-
-    K key
-    V value
-
-    Pair(K key, V value) {
-        this.key = key
-        this.value = value
-    }
-
-}
