+++ src/main/groovy/com/lakerpolling/util/Pair.groovy
@@ -1,0 +1,14 @@
+package com.lakerpolling.util
+
+
+class Pair<K, V> {
+
+    K key
+    V value
+
+    Pair(K key, V value) {
+        this.key = key
+        this.value = value
+    }
+
+}
