+++ src/test/groovy/com/lakerpolling/models/authentication/OAuthIdSpec.groovy
@@ -1,0 +1,22 @@
+package com.lakerpolling.models.authentication
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.domain.DomainClassUnitTestMixin} for usage instructions
+ */
+@TestFor(OAuthId)
+class OAuthIdSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
