+++ src/test/groovy/com/lakerpolling/services/GoogleAuthServiceSpec.groovy
@@ -1,0 +1,22 @@
+package com.lakerpolling.services
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.services.ServiceUnitTestMixin} for usage instructions
+ */
+@TestFor(GoogleAuthService)
+class GoogleAuthServiceSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
