--- src/test/groovy/com/openclicker/controllers/DashboardControllerSpec.groovy
@@ -1,22 +1,0 @@
-package com.openclicker.controllers
-
-import grails.test.mixin.TestFor
-import spock.lang.Specification
-
-/**
- * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
- */
-@TestFor(DashboardController)
-class DashboardControllerSpec extends Specification {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
