--- src/test/groovy/com/openclicker/models/authentication/OAuthIdSpec.groovy
@@ -1,22 +1,0 @@
-package com.openclicker.models.authentication
-
-import grails.test.mixin.TestFor
-import spock.lang.Specification
-
-/**
- * See the API for {@link grails.test.mixin.domain.DomainClassUnitTestMixin} for usage instructions
- */
-@TestFor(OAuthId)
-class OAuthIdSpec extends Specification {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
